more js docs
