+++ tests/npm-app/app.json
@@ -2,7 +2,7 @@
     "endpoints": {
       "/crypto": {
         "get": {
-          "js_module": "src/endpoints.js",
+          "js_module": "src/endpoints/crypto.js",
           "js_function": "crypto",
           "forwarding_required": "always",
           "execute_locally": false,
@@ -38,7 +38,7 @@
       },
       "/partition": {
         "post": {
-          "js_module": "src/endpoints.js",
+          "js_module": "src/endpoints/partition.js",
           "js_function": "partition",
           "forwarding_required": "always",
           "execute_locally": false,
@@ -80,7 +80,7 @@
       },
       "/proto": {
         "post": {
-          "js_module": "src/endpoints.js",
+          "js_module": "src/endpoints/proto.js",
           "js_function": "proto",
           "forwarding_required": "always",
           "execute_locally": false,
