+++ tests/npm-app/demo/index.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 
 <script type="module">
-import * as endpoints from "../dist/src/src/endpoints.js";
+import * as endpoints from "../dist/src/src/endpoints/all.js";
 import { CCFBody } from "./utils.js";
 
 const request = {
