+++ tests/npm-app/src/endpoints/all.ts
@@ -1,0 +1,3 @@
+export * from './crypto'
+export * from './partition'
+export * from './proto'
